청구서 금액 오버플로 방지(long/BIGINT) + 컨텍스트 기동 테스트 활성화
직전 리뷰에서 넘겨짚었던 우려들을 실제로 검증하고, 진짜 하자 2건만 수정한다.

- 금액 오버플로: 공급가/부가세/합계를 int→long, invoice 3열을 INT→BIGINT(V24).
  단가 상한(1천만)×수백 좌석이면 int(약 21.4억)를 넘겨 음수로 오버플로되던 잠재 버그.
  회귀 테스트 추가(1천만×1000명=100억을 정확히 계산).
- 컨텍스트 기동 테스트(@SpringBootTest) @Disabled 제거: 별도 CI가 없어 스킵 실익이 없고,
  DI 배선 회귀(생성자 다중정의·@Autowired 누락 — 직전 PR에서 라이브 부팅으로만 잡힌 부류)를
  mvn test에서 잡도록 상시 활성화.

검증 결과(넘겨짚음 정정): '월중 최대 누락 가능'은 오판 — 활성 전이는 초대 활성화·상태 변경
둘뿐이고 둘 다 touchSeatUsage 훅이 걸려 피크가 발생 순간 기록된다. ErrorResponse 5xx 삼킴도
오판 — 핸들러가 구체 4xx 예외만 나열해 5xx는 대상 아님. 둘 다 코드 변경 불요.

검증: mvn test 388건 그린(컨텍스트 로드 포함, Skipped 0). V1→V24 적용.

+++ src/main/java/com/attendance/pro/billing/BillingDtos.java
@@ -26,9 +26,9 @@ public final class BillingDtos {
             int freeSeats,
             int billedSeats,
             int unitPrice,
-            int subtotal,
-            int vat,
-            int total,
+            long subtotal,
+            long vat,
+            long total,
             InvoiceStatus status,
             LocalDateTime issuedAt) {
