+++ src/main/java/com/attendance/pro/billing/Invoice.java
@@ -10,8 +10,8 @@ public record Invoice(
         int freeSeats,
         int billedSeats,
         int unitPrice,
-        int subtotal,
-        int vat,
-        int total,
+        long subtotal,
+        long vat,
+        long total,
         LocalDateTime issuedAt) {
 }
